rabbitmq-s3-consumer/requirements.txt
pika==1.3.1
boto3==1.28.44
python-dotenv==1.0.0
